changed maf to better calculate based on passage of time

+++ src/main/java/com/example/smashrun/Maf/MafRange.java
@@ -1,11 +1,11 @@
 package com.example.smashrun.Maf;
 
 public class MafRange {
-    public MafRange(int minLimit, int maxLimit) {
+    public MafRange(double minLimit, double maxLimit) {
         MinLimit = minLimit;
         MaxLimit = maxLimit;
     }
 
-    public int MinLimit;
-    public int MaxLimit;
+    public double MinLimit;
+    public double MaxLimit;
 }
